backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
aiohttp==3.8.6
asyncio==3.4.3
google-generativeai==0.3.2
openai==1.3.7
anthropic==0.7.8
requests==2.31.0
pydantic==2.5.0
